+++ ec/__main__.py
@@ -23,19 +23,20 @@ def main():
     from modules.helpers import load_module
     from modules import state
     
-    state.main_module_name = 'ec.__main__'
-    
     target_path = argv.pop(0)
     
     sys.argv = sys.argv[:1] + argv # alter sys.argv so that the modules could process them
     
     if path.isfile(target_path):
-      load_module(target_path)
+      module = load_module(target_path)
+      state.main_module_name = module.__name__
       
     elif path.isdir(target_path): # launch the dir with its children as groups
       from glob import glob
       
       from ec import member # ec has to be imported in order to make the import hook work
+      
+      state.main_module_name = __name__
       
       for Module in [load_module(module_path) for module_path in glob('%s/*.py' % path.abspath(target_path)) if path.isfile(module_path)]:
         __ec_member__ = getattr(Module, '__ec_member__', None)