+++ ec/modules/hooks.py
@@ -49,14 +49,14 @@ def hookIntoImport():
     return module_name
 
   # hook into __import__ to register modules when they import ec.
-  def newImp(name, globals=None, locals=None, fromlist=None, *rest):
+  def newImp(name, globals=None, locals=None, fromlist=None, *rest, **kwargs):
     if name == EcModuleName:
       core.setActiveModule(getCallingModule())
       
-      return origImp(name, globals, locals, fromlist, *rest)
+      return origImp(name, globals, locals, fromlist, *rest, **kwargs)
     
     else:
-      imported = origImp(name, globals, locals, fromlist, *rest)
+      imported = origImp(name, globals, locals, fromlist, *rest, **kwargs)
       module_name = getModuleFullName(imported, globals)
       
       if fromlist and sys.modules.has_key('%s.%s' % (module_name, fromlist[0])): # several modules from a package are being imported in a single statement. Ex: from pkg import m1, m2